Minor refactorings (#132)
* Extract addNewParamToSearchPath from SearchView

* Change position of queryParams to be more intuitive

* Renaming customSearch to be easily understandable

* Fix typo

* one more occurence of customSearch
