+++ projects/bp-gallery/src/components/common/picture-gallery/PictureScrollGrid.tsx
@@ -18,7 +18,7 @@ const PictureScrollGrid = ({
   scrollPos,
   scrollHeight,
   hashbase,
-  customSearch = false,
+  isAllSearchActive = false,
   uploadAreaProps,
   resultPictureCallback,
   bulkOperations,
@@ -30,7 +30,7 @@ const PictureScrollGrid = ({
   scrollPos: number;
   scrollHeight: number;
   hashbase: string;
-  customSearch?: boolean;
+  isAllSearchActive?: boolean;
   uploadAreaProps?: Partial<PictureUploadAreaProps>;
   resultPictureCallback?: (pictures: number) => void;
   bulkOperations?: BulkOperation[];
@@ -44,7 +44,7 @@ const PictureScrollGrid = ({
 
   const { data, loading, error, fetchMore, refetch } = useGetPictures(
     queryParams,
-    customSearch,
+    isAllSearchActive,
     sortBy
   );
