+++ projects/bp-gallery/src/components/views/picture/sidebar/picture-info/ArchiveTagField.tsx
@@ -3,8 +3,9 @@ import React, { useCallback, useContext, useEffect, useState } from 'react';
 import { FlatArchiveTag } from '../../../../../types/additionalFlatTypes';
 import { AuthRole, useAuth } from '../../../../provider/AuthProvider';
 import { useTranslation } from 'react-i18next';
-import { addNewParamToSearchPath, SearchType } from '../../../search/SearchView';
+import { SearchType } from '../../../search/SearchView';
 import useAdvancedSearch from '../../../search/helpers/useAdvancedSearch';
+import { addNewParamToSearchPath } from '../../../search/helpers/addNewParamToSearchPath';
 
 const ArchiveTagField = ({
   archiveTag,
